Working timeout
This seems to work as long as we do not convert to promise, instead we have to subscrive to the observable and call render.
Ratpack team is working on a way to use promises.
Deleted the more concise hystrix test, because we can't use promises, we cannot test in that fashion.